requirements.txt
google-generativeai
flask
streamlit
python-dotenv
